Refactor some common list editing functionalities out to a separate file
